Update README.md
This repository provides a comprehensive AWS CloudFormation template and detailed instructions to set up a secure, scalable, and private Virtual Private Cloud (VPC) on AWS. The project is tailored for small businesses, such as law firms and other organizations, requiring secure access to internal resources. It includes an OpenVPN server for secure remote access, making it an ideal solution for enabling distributed teams to connect securely to AWS infrastructure.